-Move printList function into dragging module. -Stop import of changeStatus in index as it's not used in that script anymore. -Change cursor style on list items to represent a "movable" object.
