+++ src/dragging.js
@@ -1,3 +1,16 @@
+import changeStatus from './changeStatus.js';
+
+function printList(list) {
+  const temp = document.getElementById('Container');
+  temp.innerText = '';
+  for (let i = 0; i < list.length; i += 1) {
+    list[i].add();
+    const checkbox = document.getElementById(`${i}-checkbox`);
+    checkbox.onchange = () => { changeStatus(list, i); };
+  }
+  localStorage.clear();
+  localStorage.setItem('todoArray', JSON.stringify(list));
+}
 let droppedItem;
 let held;
 function heldItem(e) {
@@ -17,6 +30,9 @@ function dragger(arr) {
   for (let i = 0; i < arr.length; i += 1) {
     arr[i].index = i;
   }
+  printList(arr);
 }
 
-export { dragger, heldItem, droppedOn };
+export {
+  dragger, heldItem, droppedOn, printList,
+};
